+++ src/svg.js
@@ -77,7 +77,7 @@ SVG.adopt = function (node) {
   if (!node) return null
 
   // make sure a node isn't already adopted
-  if (node.instance) return node.instance
+  if (node.instance instanceof SVG.Element) return node.instance
 
   if (!(node instanceof window.SVGElement)) {
     return new SVG.HtmlNode(node)
@@ -88,12 +88,14 @@ SVG.adopt = function (node) {
 
   // adopt with element-specific settings
   if (node.nodeName === 'svg') {
-    element = node.parentNode instanceof window.SVGElement ? new SVG.Nested(node) : new SVG.Doc(node)
+    element = new SVG.Doc(node)
   } else if (node.nodeName === 'linearGradient' || node.nodeName === 'radialGradient') {
     element = new SVG.Gradient(node)
   } else if (SVG[capitalize(node.nodeName)]) {
     element = new SVG[capitalize(node.nodeName)](node)
-  } else { element = new SVG.Parent(node) }
+  } else {
+    element = new SVG.Parent(node)
+  }
 
   return element
 }